// brewlog_ui/src/app/bl-api/model/yeastStarter.ts
/**
 * brewlog.api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface YeastStarter { 
    initialCells?: number;
    dryMaltExtract?: number;
    waterLitres?: number;
    calculatedYeastCells?: number;
}